requirements.txt
aiogram>=3.0.0
python-dotenv>=1.0.0
aiohttp-socks>=0.9.0